pyproject.toml
[project]
name = "play"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nest-asyncio>=1.6.0",
    "playwright>=1.54.0",
    "pydantic-ai-slim[google]>=0.4.10",
]
